perf: use list head as top for LinkedListStack
Using the head element of the SinglyLinkedList as the top of the stack
greatly improves the performance of the operations such as inserting,
removing and peeking values.

The reason for that is the implementation of the SinglyLinkedList:
operations with the last element of the list are O(n), while operating
with the head element is O(1).